Refactored components, added helper file
