+++ src/components/layouts/header.js
@@ -1,14 +1,8 @@
-import React from 'react'
-import Logo from '../logo'
-import { Header, AuthorContainer } from '../layouts/basecss'
+import React from 'react';
+import Logo from '../shared/logo';
+import Author from '../shared/author';
+import { Header, AuthorContainer } from '../layouts/basecss';
 import profilePicture from '../../../static/profile.jpeg';
-
-const Author = ({ src, name }) => (
-  <>
-    <img style={{ borderRadius: '80px', width: '40px', height: '40px' }} src={src} alt={name} />
-    <p style={{ margin: 0, width: '150px', textAlign: 'left' }}>{`Personal Blog of ${name} in Houston, TX.`}</p>
-  </>
-)
 
 export default () => (
   <Header>
@@ -16,5 +10,6 @@ export default () => (
     <AuthorContainer>
       <Author src={profilePicture} name="Asif Kabani" />
     </AuthorContainer>
+    <div>Hello</div>
   </Header>
-)
+);
