+++ src/pages/index.js
@@ -4,57 +4,12 @@ import Layout from '../components/layouts';
 import {
   Card,
   ArticleContent,
-  Category,
   Date,
   Headline,
   Excerpt,
   More,
 } from '../components/layouts/basecss';
-
-// Color label depending on blog cateogry.
-const createLabel = category => {
-  let backgroundColor = '';
-  switch (category) {
-    case 'HTML':
-      backgroundColor = '#E44D26';
-      break;
-    case 'CSS':
-      backgroundColor = '#0070BA';
-      break;
-    case 'JavaScript':
-      backgroundColor = '#63A715';
-      break;
-
-    default:
-      break;
-  }
-  return <Category style={{ backgroundColor }}>{category}</Category>;
-};
-
-export default ({ data }) => {
-  return (
-    <Layout>
-      <ArticleContent>
-        {data.allMarkdownRemark.edges.map(({ node }) => (
-          <Link
-            to={node.fields.slug}
-            style={{ textDecoration: 'none' }}
-          >
-            <Card key={node.id}>
-              {createLabel(node.frontmatter.category)}
-              <div style={{ padding: 30 }}>
-                <Headline>{node.frontmatter.title}</Headline>
-                <Date>{node.frontmatter.date}</Date>
-                <Excerpt>{node.excerpt}</Excerpt>
-                <More />
-              </div>
-            </Card>
-          </Link>
-        ))}
-      </ArticleContent>
-    </Layout>
-  )
-}
+import { createLabel } from '../components/util';
 
 export const query = graphql`
   query {
@@ -76,4 +31,26 @@ export const query = graphql`
       }
     }
   }
-`
+`;
+
+export default ({ data }) => {
+  return (
+    <Layout>
+      <ArticleContent>
+        {data.allMarkdownRemark.edges.map(({ node }) => (
+          <Link to={node.fields.slug} style={{ textDecoration: 'none' }}>
+            <Card key={node.id}>
+              {createLabel(node.frontmatter.category)}
+              <div style={{ padding: 30 }}>
+                <Headline>{node.frontmatter.title}</Headline>
+                <Date>{node.frontmatter.date}</Date>
+                <Excerpt>{node.excerpt}</Excerpt>
+                <More />
+              </div>
+            </Card>
+          </Link>
+        ))}
+      </ArticleContent>
+    </Layout>
+  );
+};
